[AJPHR-327] Required Label (Input and textarea) (#80)
## Briefing

Refactoring: Create required label responsive to the input and text-area 

## Achievements

Input

When Input is required, I should see "required" text on the right side of the label

Remove the Optional text from the Input

When Input is not required, we only see the normal Label passed. (Any optional text should appear)

Textarea

When Textarea is required, I should see "required" text on the right side of the label

When Textarea is not required, we only see the normal Label passed. (Any optional text should appear)


https://ambush-journey.atlassian.net/jira/software/projects/AJPHR/boards/1?selectedIssue=AJPHR-327

## Screenshot



https://user-images.githubusercontent.com/116045026/233173160-240c50fb-a186-4b13-96ae-b5c3d9a7c24f.mov

![Screen Shot 2023-04-20 at 09 37 32](https://user-images.githubusercontent.com/116045026/233388175-387fa216-b065-4adc-af2b-d839f5f0843f.png)
![Screen Shot 2023-04-20 at 09 37 11](https://user-images.githubusercontent.com/116045026/233388187-06406505-4839-4d8d-ae15-d97904fbb22e.png)